:sparkles: Add the `FlowEditor` feature.
