+++ frontend/src/pages/flows/[id]/index.tsx
@@ -1,34 +1,7 @@
-import { Box, Flex, Heading } from '@chakra-ui/react';
-import { isNaN } from 'lodash';
+import { FlowEditor } from '@FlowEditor';
 import { NextPage } from 'next';
 import Error from 'next/error';
 import { useRouter } from 'next/router';
-
-import { Diagram } from '@/components/Diagram';
-import { DiagramProvider } from '@/diagrams';
-import { useBootstrapFlow } from '@/hooks/useBootstrapFlow';
-
-type Props = {
-  id: number;
-};
-
-const Component: React.VFC<Props> = ({ id }) => {
-  const { flow } = useBootstrapFlow({ id });
-
-  return (
-    <Flex direction="column" h="100vh">
-      <Box w="100%" bg="gray.600" px={10} py={6} as="nav">
-        <Heading size="md" color="white">
-          {flow?.name}
-        </Heading>
-      </Box>
-
-      <Box w="full" flex="1" maxW="full" position="relative" overflowX="auto">
-        <DiagramProvider>{flow && <Diagram flow={flow} />}</DiagramProvider>
-      </Box>
-    </Flex>
-  );
-};
 
 const Index: NextPage = () => {
   const router = useRouter();
@@ -40,7 +13,7 @@ const Index: NextPage = () => {
   const parsedId = parseInt(id);
   if (isNaN(parsedId)) return <Error statusCode={404} />;
 
-  return <Component id={parsedId} />;
+  return <FlowEditor id={parsedId} />;
 };
 
 export default Index;
